+++ yesod-core/ChangeLog.md
@@ -1,6 +1,10 @@
 # ChangeLog for yesod-core
 
-## 1.6.28.0 (unreleased)
+## 1.6.29.0
+
+* Allow users to configure whether nested routes fall through or 404. []()
+
+## 1.6.28.0
 
 * Allow users to control generation of the `resourcesSite :: [ResourceTree String]` value. [#1881](https://github.com/yesodweb/yesod/pull/1881)